==> 06_exploratory_analysis.Rmd <==
---
title: "MB1B exploratory analysis"
author: "Krista Byers-Heinlein et al"
date: "10/16/2019"
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: yes
---

# Intro

This RMarkdown file reports additional exploratory analyses of the ManyBabies 1 Bilingual (IDS Preference) dataset. 

```{r setup, echo=FALSE, message=FALSE}
source("helper/common.R")
source("helper/ma_helper.R")
library(dplyr)
library(tidyverse)
library(lme4)
library(lmerTest)

export_figs <- FALSE
```

Read datafiles and generate dataframes for analyses

```{r}
#If column types aren't specified, read_csv makes guesses about them based on the first 1000 rows of data. Below, data are read in with column types specified for columns that will be used in this .Rmd file. `col_only` in the code means that all columns that aren't specified will be dropped. (see issue #66 for more info) For now, all columns are read in as  a double (using shortcut "d"), logical ("l"), or character ("c"). 

validated_df <- read_csv("processed_data/02_validated_output.csv", col_types = cols_only(method = "c", subid = "c", lab = "c", trial_type = "c", lang_group = "c", session_error_type_recoded = "c", age_group = "c", age_days = "d", nae_exp = "d", SES = "d", age_mo = "d", stimulus_num = "d", looking_time = "d", trial_num = "d", nae = "l", trial_error = "l", session_error = "l", pilot = "l", bilingual_exposure = "l", comm_lang_learn = "l", monolingual_exposure = "l", td = "l", full_term = "l", trial_error_recoded = "l", out_of_age = "l", missing = "l", gender = "c"))

multilingual_ID <- read_csv("metadata/multilingual.csv", col_types = cols(lab = "c", subid = "c", lang1_exposure = "d", lang2_exposure = "d", lang3_exposure = "d", lang4_exposure = "d", bilingual = "l", bilingual_exposure = "l"))

bilinguals <- read_csv("processed_data/04_matched_dataset_trial.csv", col_types = cols_only(trial_num = "d", lab = "c", subid = "c", trial_type = "c", stimulus_num = "d", looking_time = "d", missing = "l", method = "c", age_group = "c", age_mo = "d", age_days = "d", lang_group = "c", nae = "l", nae_exp = "d", SES = "d", gender = "c")) %>%
  filter(lang_group == "bilingual")

d_matched <- read_csv("processed_data/04_matched_dataset_trial.csv", col_types = cols_only(trial_num = "d", lab = "c", subid = "c", trial_type = "c", stimulus_num = "d", looking_time = "d", missing = "l", method = "c", age_group = "c", age_mo = "d", age_days = "d", lang_group = "c", nae = "l", nae_exp = "d", SES = "d", gender = "c")) %>%
  mutate(lang_group = fct_relevel(lang_group, "monolingual", "bilingual"))

d_full <- read_csv("processed_data/04_full_dataset_trial.csv", col_types = cols_only(trial_num = "d", lab = "c", subid = "c", trial_type = "c", stimulus_num = "d", looking_time = "d", missing = "l", method = "c", age_group = "c", age_mo = "d", age_days = "d", lang_group = "c", nae = "l", nae_exp = "d", SES = "d", gender = "c")) %>%
  mutate(lang_group = fct_relevel(lang_group, "monolingual", "bilingual"))

d_diffs_matched_n_5 <- read_csv("processed_data/04_matched_dataset_diff_n_5.csv", col_types = cols_only(trial_num = "d", lab = "c", subid = "c", stimulus_num = "d", method = "c", age_days = "d", age_mo = "d", age_group = "c", lang_group = "c", nae = "l", ADS = "d", IDS = "d", diff = "d", nae_exp = "d", SES = "d", gender = "c")) %>%
  mutate(lang_group = fct_relevel(lang_group, "monolingual", "bilingual"))

d_diffs_full_n_5 <- read_csv("processed_data/04_full_dataset_diff_n_5.csv", col_types = cols_only(trial_num = "d", lab = "c", subid = "c", stimulus_num = "d", method = "c", age_days = "d", age_mo = "d", age_group = "c", lang_group = "c", nae = "l", ADS = "d", IDS = "d", diff = "d", nae_exp = "d", SES = "d", gender = "c")) %>%
  mutate(lang_group = fct_relevel(lang_group, "monolingual", "bilingual"))
```

Multilingual infants' dataset
```{r}
multilingual_df <- validated_df %>% filter(subid %in% multilingual_ID$subid) %>% 
  filter(full_term == TRUE, out_of_age == TRUE, session_error == FALSE, trial_error == FALSE)
  
length(unique(multilingual_df$subid))
```

LMM model 2: exploring the relationship between NAE(%) and IDS in bilinguals who were exposed to NAE. Effect still exists! :-)
```{r}
d_lmer_nae <- bilinguals %>%
  filter(trial_type != "train") %>%
  mutate(log_lt = log(looking_time),
         age_mo = scale(age_mo, scale = FALSE),
         trial_num = trial_num, 
         item = paste0(stimulus_num, trial_type)) %>%
  filter(!is.na(log_lt), !is.infinite(log_lt)) %>% 
  filter(nae_exp > 0)

nae_only_lmer <- lmer(log_lt~
  trial_type * method + age_mo * trial_num +
  trial_type * trial_num +
  trial_type * age_mo * nae_exp +
  (1 | lab) +  
  (1 |subid) +
  (1 | item),
  data = d_lmer_nae)
 
summary(nae_only_lmer) 

kable(summary(nae_only_lmer)$coefficients, digits = 3)
```

# Re-run the mixed-effect models by treating age as dummy variable
## Matched dataset (center age by treating it as a dummy variable)

LMM model 1 (IDS pref) 
Matched dataset
```{r}
d_lmer_matched_dummy <- d_matched %>%
  filter(trial_type != "train") %>%
  mutate(log_lt = log(looking_time),
         age_group = as.factor(age_group),
         trial_num = trial_num, 
         item = paste0(stimulus_num, trial_type)) %>%
  filter(!is.na(log_lt), !is.infinite(log_lt))

mod_lmer_matched_dummy <- lmer(log_lt ~
  trial_type * method + age_group * trial_num +
  trial_type * trial_num +
  trial_type * age_group * nae +
  trial_type * age_group * lang_group +
  (1 |subid) +
  (lang_group |lab) + 
  (1 |item),
  data = d_lmer_matched_dummy)

summary(mod_lmer_matched_dummy) 

kable(summary(mod_lmer_matched_dummy)$coefficients, digits = 3)
```

Full dataset
```{r}
d_lmer_unmatched_dummy <- d_full %>%
  filter(trial_type != "train") %>%
  mutate(log_lt = log(looking_time),
         age_group = as.factor(age_group),
         trial_num = trial_num, 
         item = paste0(stimulus_num, trial_type)) %>%
  filter(!is.na(log_lt), !is.infinite(log_lt))

mod_lmer_unmatched_dummy <- lmer(log_lt~
  trial_type * method + age_group * trial_num +
  trial_type * trial_num +
  trial_type * age_group * nae +
  trial_type * age_group * lang_group +
  (1 |subid) +
  (1 |lab) +
  (1 |item), 
  data = d_lmer_unmatched_dummy)

summary(mod_lmer_unmatched_dummy) 

kable(summary(mod_lmer_unmatched_dummy)$coefficients, digits = 3)
```

LMM model 2 (nae_exp) 
```{r}
d_lmer_matched_Lmm2_dummy <- d_lmer_matched_dummy %>% filter(lang_group=="bilingual")

nae_lmer_matched_dummy <- lmer(log_lt~
  trial_type * method + age_group * trial_num +
  trial_type * trial_num +
  trial_type * age_group * nae_exp +
  (1 |subid) +
  (trial_type |lab) +
  (1 |item), 
  data = d_lmer_matched_Lmm2_dummy)
 
summary(nae_lmer_matched_dummy) 

kable(summary(nae_lmer_matched_dummy)$coefficients, digits = 3)
```

LMM model 3 (SES)
Matched dataset
```{r}
d_lmer_matched_SES_dummy <- d_matched %>% 
  filter(!is.na(SES)) %>% 
  mutate(log_lt = log(looking_time),
         age_group = as.factor(age_group),
         SES_c = scale(SES, scale = FALSE))

SES_lmer_matched_dummy <- lmer(
  log_lt ~
  trial_type * method +
  age_group * trial_num +
  trial_type * trial_num +
  trial_type * age_group * nae +
  trial_type * age_group * lang_group +
  trial_type * age_group * SES_c +
  (1 | subid) +
  (1 | lab),
  data = d_lmer_matched_SES_dummy
  )
 
summary(SES_lmer_matched_dummy) 

kable(summary(SES_lmer_matched_dummy)$coefficients, digits = 3)
```

Full dataset
```{r}
d_lmer_unmatched_SES_dummy <- d_full %>% 
  filter(!is.na(SES)) %>% 
  mutate(log_lt = log(looking_time),
         age_group = as.factor(age_group),
         SES_c = scale(SES, scale = FALSE), 
         item = paste0(stimulus_num, trial_type))


SES_lmer_unmatched_dummy <- lmer(
  log_lt ~
  trial_type * method +
  age_group * trial_num +
  trial_type * trial_num +
  trial_type * age_group * nae +
  trial_type * age_group * lang_group +
  trial_type * age_group * SES_c +
  (1 | subid) +
  (1 | lab) +
  (1 | item),
  data = d_lmer_unmatched_SES_dummy
  )

summary(SES_lmer_unmatched_dummy)

kable(summary(SES_lmer_unmatched_dummy)$coefficients, digits = 3)
```

# Re-run meta-analysis using n >= 5 criteria
### Matched dataset
```{r}
source("helper/ma_helper.R")

ages <- d_matched %>%
  group_by(lab, age_group, lang_group, subid) %>%
  summarise(age_mo = mean(age_mo)) %>%
  summarise(age_mo = mean(age_mo)) %>% 
  mutate(age_mo_centered = scale(age_mo, scale = FALSE))

ds_zt_matched_set_n_5 <- d_diffs_matched_n_5 %>%
  group_by(lab, age_group, lang_group, subid) %>%
  summarise(d = mean(diff, na.rm = TRUE)) %>%
  group_by(lab, age_group, lang_group) %>%
  summarise(mean_d = mean(d, na.rm=TRUE),
            d_z = mean(d, na.rm = TRUE) / sd(d, na.rm = TRUE),
            n = length(unique(subid)),
            d_z_var = d_var_calc(n, d_z)) %>%
  filter(n >= 5) %>%
  left_join(ages) %>%
  filter(!is.na(d_z)) # CHECK THIS


```

### Unmatched dataset

```{r}
ages_full <- d_full %>%
  group_by(lab, age_group, lang_group, subid) %>%
  summarise(age_mo = mean(age_mo)) %>%
  summarise(age_mo = mean(age_mo)) %>% 
  mutate(age_mo_centered = scale(age_mo, scale = FALSE))

ds_zt_unmatched_set_n_5 <- d_diffs_full_n_5 %>%
  group_by(lab, age_group, lang_group, subid) %>%
  summarise(d = mean(diff, na.rm = TRUE)) %>%
  group_by(lab, age_group, lang_group) %>%
  summarise(mean_d=mean(d,na.rm=TRUE),
            d_z = mean(d, na.rm = TRUE) / sd(d, na.rm = TRUE),
            n = length(unique(subid)),
            d_z_var = d_var_calc(n, d_z)) %>%
  filter(n >= 5) %>%
  left_join(ages_full) %>%
  filter(!is.na(d_z)) # CHECK THIS
```

## Effect size-based meta-analysis

First, interecept-only model: examine the mean effect size across all groups in MB1B, regardless of language backgrounds.

### The average IDS effect size in the Matched dataset

```{r}
intercept_matched_set_n_5 <- metafor::rma(d_z ~ 1, vi = d_z_var, slab = lab, data = 
                                        ds_zt_matched_set_n_5, method = "REML")
summary(intercept_matched_set_n_5)
```

### The average IDS effect size in the Unmatched dataset

```{r}
intercept_unmatched_set_n_5 <- metafor::rma(d_z ~1, vi = d_z_var, slab = lab, data = ds_zt_unmatched_set_n_5, method = "REML")
summary(intercept_unmatched_set_n_5)
```

Note that in both matched and unmatched models, Q statistics indicate that there is non-significant heterogenenity. Here, we report the unmatched model. Any numbers in $I^2$ = `r intercept_unmatched_set$I2` is essentially meaningless. Reason for I not being zero when Q statistics is non-significant is as follow: $I^2 = (Q - df) / Q$ (following [Higgins et al., 2003](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC192859/)). When Q drops below df, as it does here (Q = `r intercept_unmatched_set$Q` and df = `r nrow(ds_zt_unmatched_set) -1`), then $I^2$ is defined as being zero, e.g. no heterogeneity. 

## Language*age moderation (pre-reg)

### Matched dataset

```{r}
lang_mod_matched_n_5 <- metafor::rma(d_z ~ age_mo_centered*lang_group, vi = d_z_var,
                    slab = lab, data = ds_zt_matched_set_n_5, method = "REML")
summary(lang_mod_matched_n_5)
```
The lang_only_matched model has suggested that there is no significant difference in IDS preference between monolinguals and bilinguals
```{r}
lang_only_matched_n_5 <- metafor::rma(d_z ~ lang_group, vi = d_z_var,
                    slab = lab, data = ds_zt_matched_set_n_5, method = "REML")
summary(lang_only_matched_n_5)
```
But as indicated in the RR analysis plan, we will report the average effect size for each language group
```{r}
#monolinguals' IDS ES in the matched dataset
mono_sample_matched_n_5 <- ds_zt_matched_set_n_5 %>% filter(lang_group == "monolingual") 

mono_only_matched_n_5 <- metafor::rma(d_z ~ 1, vi = d_z_var, data = mono_sample_matched_n_5, slab = lab, method = "REML")
summary(mono_only_matched_n_5)

#bilinguals' IDS ES in the matched dataset
bil_sample_matched_n_5 <- ds_zt_matched_set_n_5 %>% filter(lang_group == "bilingual") 

bil_only_matched_n_5 <- metafor::rma(d_z ~ 1, vi = d_z_var, data = bil_sample_matched_n_5, slab = lab, method = "REML")
summary(bil_only_matched_n_5)
```


### Unmatched dataset

```{r}
lang_mod_unmatched_n_5 <- metafor::rma(d_z ~ age_mo_centered*lang_group, vi = d_z_var,
                    slab = lab, data = ds_zt_unmatched_set_n_5, method = "REML")
summary(lang_mod_unmatched_n_5)
```
Neither age nor language group (plus their interaction term) is significant

We will report the average effect size for each language group
```{r}
#monolinguals' IDS ES in the unmatched dataset
mono_sample_unmatched_n_5 <- ds_zt_unmatched_set_n_5 %>% filter(lang_group == "monolingual") 

mono_only_unmatched_n_5 <- metafor::rma(d_z ~ 1, vi = d_z_var, data = mono_sample_unmatched_n_5, slab = lab, method = "REML")
summary(mono_only_unmatched_n_5)

#bilinguals' IDS ES in the unmatched dataset 
bil_sample_unmatched_n_5 <- ds_zt_unmatched_set_n_5 %>% filter(lang_group == "bilingual") 

bil_only_unmatched_n_5 <- metafor::rma(d_z ~ 1, vi = d_z_var, data = bil_sample_unmatched_n_5, slab = lab, method = "REML")
summary(bil_only_unmatched_n_5)
```
